// packages/react-packages/toast/src/constants/enums.ts
export enum ToastPosition {
  TopLeft = 'top-left',
  TopCenter = 'top-center',
  TopRight = 'top-right',
  BottomLeft = 'bottom-left',
  BottomCenter = 'bottom-center',
  BottomRight = 'bottom-right',
}

export enum ToastType {
  Success = 'success',
  Error = 'error',
  Info = 'informative',
  Warning = 'warning',
  /* react-hot-toast types
  Loading = 'loading',
  Blank = 'blank',
  Custom = 'custom',
  */
}